Security

+++ src/Controller/CategoryController.php
@@ -91,25 +91,34 @@ class CategoryController extends AbstractController
     )]
     public function create(Request $request): Response
     {
-        $category = new Category();
-        $form = $this->createForm(CategoryType::class, $category);
-        $form->handleRequest($request);
+        if ($this->isGranted('ROLE_ADMIN')) {
+            $category = new Category();
+            $form = $this->createForm(CategoryType::class, $category);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->categoryService->save($category);
+            if ($form->isSubmitted() && $form->isValid()) {
+                $this->categoryService->save($category);
 
+                $this->addFlash(
+                    'success',
+                    $this->translator->trans('message.created_successfully')
+                );
+
+                return $this->redirectToRoute('category_index');
+            }
+
+            return $this->render(
+                'category/create.html.twig',
+                ['form' => $form->createView()]
+            );
+        } else {
             $this->addFlash(
-                'success',
-                $this->translator->trans('message.created_successfully')
+                'danger',
+                $this->translator->trans('message.access_denied')
             );
 
             return $this->redirectToRoute('category_index');
         }
-
-        return $this->render(
-            'category/create.html.twig',
-            ['form' => $form->createView()]
-        );
     }
 
     /**
@@ -123,34 +132,43 @@ class CategoryController extends AbstractController
     #[Route('/{id}/edit', name: 'category_edit', requirements: ['id' => '[1-9]\d*'], methods: 'GET|PUT')]
     public function edit(Request $request, Category $category): Response
     {
-        $form = $this->createForm(
-            CategoryType::class,
-            $category,
-            [
-                'method' => 'PUT',
-                'action' => $this->generateUrl('category_edit', ['id' => $category->getId()]),
-            ]
-        );
-        $form->handleRequest($request);
+        if ($this->isGranted('ROLE_ADMIN')) {
+            $form = $this->createForm(
+                CategoryType::class,
+                $category,
+                [
+                    'method' => 'PUT',
+                    'action' => $this->generateUrl('category_edit', ['id' => $category->getId()]),
+                ]
+            );
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->categoryService->save($category);
+            if ($form->isSubmitted() && $form->isValid()) {
+                $this->categoryService->save($category);
 
+                $this->addFlash(
+                    'success',
+                    $this->translator->trans('message.edited_successfully')
+                );
+
+                return $this->redirectToRoute('category_index');
+            }
+
+            return $this->render(
+                'category/edit.html.twig',
+                [
+                    'form' => $form->createView(),
+                    'category' => $category,
+                ]
+            );
+        } else {
             $this->addFlash(
-                'success',
-                $this->translator->trans('message.edited_successfully')
+                'danger',
+                $this->translator->trans('message.access_denied')
             );
 
             return $this->redirectToRoute('category_index');
         }
-
-        return $this->render(
-            'category/edit.html.twig',
-            [
-                'form' => $form->createView(),
-                'category' => $category,
-            ]
-        );
     }
 
     /**
@@ -164,42 +182,51 @@ class CategoryController extends AbstractController
     #[Route('/{id}/delete', name: 'category_delete', requirements: ['id' => '[1-9]\d*'], methods: 'GET|DELETE')]
     public function delete(Request $request, Category $category): Response
     {
-        if(!$this->categoryService->canBeDeleted($category)) {
+        if ($this->isGranted('ROLE_ADMIN')) {
+            if (!$this->categoryService->canBeDeleted($category)) {
+                $this->addFlash(
+                    'warning',
+                    $this->translator->trans('message.category_contains_recipes')
+                );
+
+                return $this->redirectToRoute('category_index');
+            }
+
+            $form = $this->createForm(
+                CategoryType::class,
+                $category,
+                [
+                    'method' => 'DELETE',
+                    'action' => $this->generateUrl('category_delete', ['id' => $category->getId()]),
+                ]
+            );
+            $form->handleRequest($request);
+
+            if ($form->isSubmitted() && $form->isValid()) {
+                $this->categoryService->delete($category);
+
+                $this->addFlash(
+                    'success',
+                    $this->translator->trans('message.deleted_successfully')
+                );
+
+                return $this->redirectToRoute('category_index');
+            }
+
+            return $this->render(
+                'category/delete.html.twig',
+                [
+                    'form' => $form->createView(),
+                    'category' => $category,
+                ]
+            );
+        } else {
             $this->addFlash(
-                'warning',
-                $this->translator->trans('message.category_contains_recipes')
+                'danger',
+                $this->translator->trans('message.access_denied')
             );
 
             return $this->redirectToRoute('category_index');
         }
-
-        $form = $this->createForm(
-            CategoryType::class,
-            $category,
-            [
-                'method' => 'DELETE',
-                'action' => $this->generateUrl('category_delete', ['id' => $category->getId()]),
-            ]
-        );
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->categoryService->delete($category);
-
-            $this->addFlash(
-                'success',
-                $this->translator->trans('message.deleted_successfully')
-            );
-
-            return $this->redirectToRoute('category_index');
-        }
-
-        return $this->render(
-            'category/delete.html.twig',
-            [
-                'form' => $form->createView(),
-                'category' => $category,
-            ]
-        );
     }
 }
